mv `test_universal_assisted_generation.py` to `generation/test_candidate_generator.py`
